refactor: logout and login moved to components
add login function to component
remove logout button from profil component and add it to navigation bar
