+++ app/routes/login.js
@@ -2,10 +2,6 @@ const express = require("express");
 const router = express.Router();
 const authenticateUser = require("./middlewares/authenticateUser");
 
-router.get("/login", authenticateUser, (req, res) => {
-  console.log("test");
-});
-
 router.post("/", authenticateUser, (req, res, next) => {
   console.log("Logged in as", req.user);
   res.send(req.user);
