+++ frontend/package.json
@@ -10,6 +10,7 @@
     "bootstrap": "^5.2.2",
     "http-proxy-middleware": "^2.0.6",
     "react": "^18.2.0",
+    "react-bootstrap": "^2.5.0",
     "react-client-session": "^0.0.8",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.2",
