+++ frontend/src/index.js
@@ -5,7 +5,6 @@ import App from "./App";
 import { BrowserRouter as Router } from "react-router-dom";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 import reportWebVitals from "./reportWebVitals";
-import { AuthProvider } from "./auth/useAuth";
 
 import "bootstrap/dist/css/bootstrap.min.css";
 
@@ -13,9 +12,7 @@ const rootElement = document.getElementById("root");
 const root = createRoot(rootElement);
 root.render(
   <Router>
-    <AuthProvider>
-      <App />
-    </AuthProvider>
+    <App />
   </Router>
 );
 serviceWorkerRegistration.register();
